Refactor the Queue tests
When making some changes to Queue (coming up in a future commit) I found
that these tests were hard to work with. There's a lot of clever and
unique interdependent fixtures that're shared between all the tests,
which makes the tests hard to maintain. Adding a new test is
particularly difficult as the new test ought to make use of the
fixture-complex like the existing tests do, but writing such a new test
requires understanding the fixture-complex and possibly adapting it to
the needs of the new test, without breaking any of the existing tests.

I've changed the tests in a way that should make maintaining them, and
adding new tests, much easier. This makes the new test that I'm adding
in a future commit much easier to write.

I think these tests have fallen into a typical difficult tests pitfall:

Pitfall 1: Lots of complex inter-related fixtures
-------------------------------------------------

These tests have a bunch of complex and inter-related fixtures and
constants:

* `LIMIT` is `2` as a constant. All the tests pass it as the `limit`
  argument to `Queue.sync()` and various of the fixtures below also use it

* The `annotations()` fixture adds `LIMIT + 1` annotations to the DB and
  returns them

* The `all_annotation_ids` fixture depends on the `annotations()`
  fixture and returns a list of all their IDs

* The `annotation_ids` fixture depends on the `annotations()` fixture
  and returns a list of all the IDs of the first `LIMIT` annotations

* The `index()` fixture returns a functions that indexes the given
  annotations into Elasticsearch.

* The `with_indexed_annotations()` fixture depends on the `annotations`
  fixture and indexes all the annotations into Elasticsearch

* The `add_all()` fixture that returns a function that, given no
  arguments(!), takes the annotation ID from the `annotation_ids` fixture
  and adds `sync_annotation` jobs for them all to the queue.

* The `with_queued_annotations()` fixture depends on the `add_all()`
  fixture and calls it, so it adds `sync_annotation` jobs for all the
  annotations from the `annotations` fixture to the job queue.

Pitfall 2: Tests tightly coupled to each other via fixtures
-----------------------------------------------------------

All the tests use all the fixtures above, so you can't make any changes
to any of the fixtures without potentially breaking some of the existing
tests. And in some cases you might have broken the test such that it
always passes even if the code is wrong, and it'd be hard to know that
you had done this.

Pitfall 3: Unusual fixtures
---------------------------

Wherever possible it's best to use common test patterns. For example we
often write patch fixtures, or use test factories. Services tests often
have a fixture called `svc` that returns the service under test. We use
the `pyramid_request` fixture all over the place. Etc etc.

Common patterns like this are easier to understand because they're
familiar. You learn them once and then you recognise them everywhere.

The fixtures in these tests are very unique to these tests, so you have
to read through and understand them all just to understand these tests.

Pitfall 4: Hiding too much stuff in fixtures
--------------------------------------------

The tests are just hard to understand because so much is done in
fixtures:

    def test_it_ignores_jobs_that_arent_scheduled_yet(
        self, add_all, batch_indexer, queue
    ):
        add_all(schedule_in=ONE_WEEK_IN_SECONDS)

        counts = queue.sync(LIMIT)

        assert counts == {}
        batch_indexer.index.assert_not_called()

The call to `add_all()` causes 3 annotations to be added to the DB but
does not add those annotations to Elasticsearch and it adds
`sync_annotation` jobs to the queue for each of them. But this is all
totally invisible in the test. To understand it you'd need to know about
the `LIMIT` constant, the `annotations` and `annotation_ids` fixtures,
and the `queue` fixture, as well as the `add_all()` fixture itself.

Solution
--------

I've gotten rid of `LIMIT` and the `annotations()`,
`all_annotation_ids()`, `annotation_ids()`,
`with_indexed_annotations()`, `with_queued_annotations()`, and
`add_all()` fixtures, and replaced them with one fixture that follows a
common pattern: a `SyncAnnotationJobFactory()`.

This is a test factory that's based on the generic `factories.Job()` but
that specifically creates `sync_annotation` jobs. It will automatically
create an annotation for each job, and supports creating unsheduled
jobs, expired jobs, multiple jobs, `force=True` jobs, and either
indexing the annotation into Elasticsearch or not. Its usage is familiar
from how you'd use any test factory.

All the tests are changed to use this factory, and are much clearer as a
result:

    def test_it_ignores_jobs_that_arent_scheduled_yet(
        self, batch_indexer, now, queue, SyncAnnotationJobFactory
    ):
        SyncAnnotationJobFactory(scheduled_at=now + datetime_.timedelta(hours=1))

        counts = queue.sync(1)

        assert counts == {}
        batch_indexer.index.assert_not_called()

You can see the test set up at the top of the test method itself,
instead of having to infer it from a bunch of interdependent shared
fixtures below.

Given familiarity with test factories it's pretty obvious what
`SyncAnnotationJobFactory()` without looking but if you want to you can
scroll down to his one fixture and read its docstring.