Simplify SyncAnnotationJobFactory
I think the code is actually better if this factory is *less* clever and
doesn't support index the job's annotation into Elasticsearch for you.

Tests now have to do this:

    annotation = factories.Annotation()
    index(annotation)
    factories.SyncAnnotationJob(annotation=annotation)

Whereas previously, at least if the test didn't need access to
`annotation` it could be a one-liner:

    SyncAnnotationJobFactory(index=True)

But I think the three-line version is arguably better because it makes
it clearer in the test code what is happening: 1. An annotation is being
created; 2. The annotation is being indexed; 3. A "sync_annotation" job
for the annotation is being created. Rather than the reader having to
know about magic that happens in SyncAnnotationJobFactory to create an
annotation and index it.

And the new factories.SyncAnnotationJob class is much simpler than
SyncAnnotationJobFactory, with less use of factory_boy magic. It doesn't
have to take a special `index` argument and override _create().

Also, since the factory no longer depends on the `index()` fixture, it
no longer needs to be defined as an inner-class inside a pytest fixture
method. It can just be defined like a normal factory class.