Replace Stale app with GitHub Action
Replace the "Stale" GitHub App that we use to label and automatically
close inactive pull requests with a GitHub Action.

We encountered an issue where the app was not closing stale issues
despite being inactive for 30+ days. The bot's configuration doesn't
specify how long to wait but according to its README [1] this should
happen after 7 days by default. Looking at the state of the issue
tracker it doesn't appear to be actively maintained. Therefore we're
going to try the Stale action from the official GitHub Actions org
instead [2].

[1] https://github.com/probot/stale
[2] https://github.com/actions/stale