Enable listening to localhost IP
Currently, h only listens and responds to localhost
(`http://localhost:3000`).

Is there any security risk by enabling h to listen and respond to
localhost's IP? `http://<IP localhost>:3000`

Changing the host to `0.0.0.0` is necessary to debug the client on
certain situation (see step 3 of the `preparation` section
[here](https://stackoverflow.com/c/hypothesis/questions/388/389#389)).

If changing this setting is not a risk, then it would facilitate
debugging the client and developers of the client would appreciate it.