Close Elasticsearch TCP connection after each test that uses it
In Python 3.9.x, search tests were failing with a warning:

```
ResourceWarning: unclosed <socket.socket ...
```

Fix the issue by closing the connection to Elasticsearch after each test
that uses the `es_client` fixture.

In the latest version of the Elasticsearch Python library the main
`Elasticsearch` class (used as `es_client.conn`) has a `close` method
for this. The version we are using does not, so we have to call `close`
on the underlying transport instead.