Fix test failure in h.schema.forms.admin.group under Python 3.9
The test expects parsing `foo:123` as a URL to result in a URL with no
scheme or host. `urlsplit` produces different results for this in
Python 3.6 and Python 3.9. Under Python 3.6 `foo:123` is parsed as a URL
with only a path. Under Python 3.9 this is parsed as a URL with scheme
`foo` and path `123`.

Fix the issue by making the validation in the `parse_origin` utility
more strict so that both a scheme and a netloc are required.