Link to client boot script directly in /app.html
Link to the client boot script (https://cdn.hypothes.is/hypothesis)
directly in the client's sidebar HTML page (/app.html) as well as the
/notebook and /a/{id} routes, instead of referencing the /embed.js route
which redirects to the client boot script.

Referencing /embed.js instead of the client boot script historically did not
incur a performance penalty in the sidebar because the redirect would
have been cached from when the client was loaded in the host page or PDF
viewer.

However there are various scenarios when this redirect is not cached,
and for these we can speed up client boot by linking directly to the
boot script. These scenarios include:

1. When the user visits the /notebook or /a/{id} h URLs in a top-level
   frame
2. In a recent browser that partitions the HTTP cache by origin,
   preventing a cached redirect from being used. Chrome started doing
   this in v86 (see
   https://developers.google.com/web/updates/2020/10/http-cache-partitioning
   and in particular [this
   section](https://developers.google.com/web/updates/2020/10/http-cache-partitioning#how_will_cache_partitioning_affect_chromes_http_cache))
3. When browser dev tools are open and caching is disabled. In this case
   the missing /embed.js => client boot script caching adds a misleading
   delay to app startup.