Replace `hkdf` with `pycryptodomex`
This replaces `hkdf` with `pycryptodomex` which is active and supports
Python 3.9. This causes no trouble with our build process (unlike
`cryptography` for the moment).

We are removing `hkdf` as it only advertises support for Python 3.4.

An example implementation using `cryptography` has been added as a
comment in case we need / want to switch one day.