Switch tests to run in headless Chrome
Switch frontend tests to run in a modern browser which better matches
the actual environment in which code is run.

There are 5 tests which currently fail when run in Chrome. These have
been temporarily marked as skipped. I would have preferred to mark them as
expected failures, but Mocha doesn't support that.