Remove gulp-newer stream when processing vendor CSS
We tried to upgrade the version of `bootstrap.css`, however the newer
version of this file was not copy over `build/styles`. This is because
(at least on mac) different version of `bootstrap.css` shows the same
date (26 Oct 1985).

```
% ls -lh ./node_modules/bootstrap/dist/css/bootstrap.css
-rw-r--r--  1 esanzgar  staff   191K 26 Oct  1985 ./node_modules/bootstrap/dist/css/bootstrap.css <- this is version 5.0

% ls -lh build/styles/bootstrap.css
-rw-r--r--  1 esanzgar  staff   193K 26 Oct  1985 build/styles/bootstrap.css <- this is version 4.6
```

I have just confirmed that `npm` publishes packages with all the
(contained) file timestamps locked to 26 Oct 1985:

https://github.com/npm/npm/issues/20439

Given the above information it is better to not rely on the timestamp.