Fix groups icon rendering in production builds
The `groups.svg` icon has a structure like this:

```
...
<g style="fill:none;...">
  <g fill="currentColor" ...>
    ...
  </g>
</g>
```

The newest version of svgo, introduced in
https://github.com/hypothesis/h/pull/6739, combines the two `<g>`
elements:

```
<g fill="currentColor" style="fill:none;...">
```

The result is that the icon ends up being transparent instead of
inherting the current CSS `color` value.

This commit fixes the issue by just removing the conflicting `style`
property. No other icons are affected AFAICS.