Update `logo.png` to current logo
`logo.png` was using an out-of-date logo. This image file was formerly
referenced as the OpenGraph image for sharing. It is no referenced used
anywhere within this codebase, and may perhaps be removed entirely once
it's verified that it is not used externally.

Note that this image now has smaller resolution, as I cannot find a
variant of the current logo in that large of resolution.