Bump h-assets to v1.0.4
This release makes cache-busting query strings on asset URLs optional.
This fixes a problem on the /docs/help page where the icon font would
fail to load because it was referenced in icomoon.css without a cache
buster. It also fixes errors in the browser console due to JS and CSS
bundles referencing sourcemaps without cache busters.