Rewrite the tests to state the behavior instead of mocking
This also removes a hatch only used by the tests. We can mock the check
function if we want to... but I'm not sure we do.

This generally stops relying on mocking the underlying object from
pyramid. This means we can refactor the code to remove that object
and we can test if we still do the same thing.