Strip Google Drive `resourcekey` parameter when normalizing URIs
The `resourcekey` query param was added to a small proportion of non-native
(eg. PDF) download and share URLs in Google Drive. It is only a
security key and does not affect which file the URL refers to.

In general adding new parameters to the query parameter blocklist used
by URI normalization is problematic (see https://github.com/hypothesis/h/issues/6552).
However this parameter was added only quite recently and almost all
affected URIs are PDFs, so the fingerprint will connect any existing
annotations made on URIs with/without the `resourcekey` parameter.