Add logging as well as a warning, and a way to trigger a similar situation
We we're attempting to use a warning to show when the permissions systems
diverged, but I'm not convinced they show up. This offers a temporary way
of triggering the same situation in live on demand to prove it.

Locally I don't see the warning, but I do see the log message.