Use yarn instead of npm for package management
We use Yarn for all of our other major frontend projects. The Yarn
lockfile format is also much more compact and diffs of Yarn lockfiles are a
lot more amenable to manual inspection.

In the process the check-dependencies dependency was removed. IIRC this
was originally aded to speed up no-op `npm install`s. A no-op `yarn install` is
fast enough that it isn't necessary.