Prepare for Elasticsearch 6.8 and 7.x migrations
Prepare for migrating from Elasticsearch 6.2 to ES 6.8 and then to ES
7.x by making h support each of these versions. The two changes that
affect h are the removal of mapping types [1] and a change to the `hits.total`
field of search responses.

Mapping types are referenced in h when creating the search index and
when indexing new documents. In ES 6.x there can only be one mapping
type in an index and it can have any name. We use "annotation". In ES
7.x the mapping type is optional but if present it must be "_doc".

These changes are handled as follows:

 - When creating indexes, use the mapping type "annotation" if the server
   version is ES 6.x or omit it for an ES 7.x server.

 - When indexing documents, include the mapping type "annotation" if the server
   version is an ES 6.x version or omit it for ES 7.x versions.

   This change means that the indexing code will work using indexes originally
   created using ES 6.x or 7.x.

 - Support parsing the different `hits.total` field formats returned by search
   responses from ES 6.x and ES 7.x.

[1] https://www.elastic.co/guide/en/elasticsearch/reference/7.17/removal-of-types.html