Add html_side_by_side feature flag
When enabled, this will cause the client to adjust the layout of web
pages when the sidebar is open, in order to display the page's content
alongside rather than under the sidebar. This feature does not work on
all pages, and the client uses heuristics to determine whether to
perform this layout adjustment or not.