Increased default timeout: tasks.yml
Changed default timeout to 900 seconds. This value is used in two
contexts:

It is passed to AWS when requesting an ODIC token. The timeout defines
the life of the token, and the default is 900 seconds. If you pass a
value lower it will result in an error.

The value is also used to constrain the length of the H task being
executed.