Remove the separate WebSocket Docker image
Our new GitHub Actions-based deployment workflows aren't building and
releasing the hypothesis-ws Docker image. The latest version of that
Docker image to be published is two months old and the workflows are
just repeatedly deploying that old Docker image to the Elastic Beanstalk
environment.

Unlike our other application repos the h repo has two separate
`Dockerfile`'s that need to be built, released and deployed as separate
Docker images/containers running in separate Elastic Beanstalk
environments.

This PR removes the second `Dockerfile` so that h is consistent with the
rest of our projects in having just one Docker image and will now work
with our release and deployment workflow. The h repo also gets much
simpler as a result. The ability to run only the WebSocket process in a
separate Elastic Beanstalk environment is maintained by using
environment variables to tell Supervisor to only start one process or
the other (web or WebSocket) depending on the environment.