Untangle the WebSocket from the internal NGINX
Untangle the WebSocket process from h's internal NGINX instance.

The idea is that h's internal NGINX will only be used for the web
process, never for the WebSocket process. The web process will always
run behind the internal NGINX (unless we one day manage to obsolete the
internal NGINX and remove it entirely). The WebSocket process will never
run behind the internal NGINX in any environment.

The Docker image exposes NGINX on port `5000` with the web process
behind NGINX, and exposes the WebSocket process on port `5001` (not
behind NGINX).

Ops will have two choices about how to deploy the WebSocket process:

1. Ops could run the web and WebSocket processes together in one
   environment. In this case we'd have to use Elastic Beanstalk's NGINX
   config (in the `deployment` repo) to route `/ws` requests to `5001`
   and other requests to `5000`. Something similar to the routing that
   this commit removes from h's internal `nginx.conf` would have to be
   added to the Elastic Beanstalk NGINX config in the `deployment` repo.

2. Ops could run the web and WebSocket processes in two separate
   environments. In the web environment port `5000` would be mapped to
   `443` whereas in the WebSocket environment `5001` would be mapped to
   `443`.

It's nice that this commit makes the h repo and its Docker image
agnostic about whether the web and WebSocket processes are deployed in a
single environment or in separate environments: that's purely an ops
concern.

Removing this ops concern from the h repo enables this commit to
simplify the h repo in a few ways:

1. The two separate `websocket-monolithic.ini` and
   `websocket-separate.ini` config files are replaced by a single
   `websocket.ini`.

2. The `WEBSOCKET_CONFIG` envvar is removed.

3. The WebSocket support is removed from `nginx.conf`.

A couple of ops changes would be needed before we could deploy this:

1. The Canadian `h-prod` environment currently relies on the WebSocket
   support in h's internal `nginx.conf`, which this commit removes. The
   Canadian environment would need to be changed to no longer rely on
   this, for example by deploying a separate `h-websocket-prod`
   environment like we have in the US.

2. The US `h-websocket-qa` and `h-websocket-prod` environments would
   need to be changed to expect the WebSocket on `5001` not `5000`.