Don't run workflows unnecessarily
Use `paths-ignore` to skip running the `ci.yml` and `deploy.yml`
workflows for pull requests where the files modified make it unnecessary
to run the workflow. For example: don't run the tests on a PR that only
modifies the documentation, or don't deploy a PR that only modifies
development dependencies.

See:

* https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#example-excluding-paths
* https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#filter-pattern-cheat-sheet