Add a service for migrating URLs
Add a service that can perform batch migration of annotations from one
set of URLs / documents to another. The initial use case is for
migrating annotations on VitalSource ebook annotations from individual
chapter URLs to whole-book URLs as part of https://github.com/hypothesis/h/issues/7709.

The migration reuses the `h.storage.update_annotation` function that was
originally used for handling annotation updates via the API, to ensure
that annotations and documents are updated in a way that is consistent
how they would be updated if users "moved" the annotations via API calls
for each annotation.