Set rate limits for URL migration tasks
These are conservative to err on the side of avoiding service disruption, yet
enough that it shouldn't take too long to migrate 14K VitalSource annotations
over ~600 URLs. Note that for the task that moves annotations, each task
processes a batch of 50, so the rate is 20*50 = 1000 per worker per minute.

The global `worker_disable_rate_limits` option takes precedence over
rate limits for individual tasks [1]. That has been enabled ever since
Celery was added to H. In order for per-task rate limits to take effect,
we need to disable it. The default rate limit for tasks is already no
limit [2].

[1] https://docs.celeryq.dev/en/stable/userguide/configuration.html#std-setting-worker_disable_rate_limits
[2] https://docs.celeryq.dev/en/stable/userguide/configuration.html#task-default-rate-limit