Lower rate limit for URL migration
During this morning's migration of ~15,600 VitalSource annotations, tasks from
the URL migration saturated available Celery workers. The best way to solve this
would likely involve leveraging the job queue more to manage processing, as we
do for annotation syncing. As a quick and easy improvement though, lower the
rate limit.