Fix an inefficient bulk annotations query
When testing the new email digests feature in production we ran into a
problem with bulk annotations API requests timing out and overloading
the DB's CPU: [Slack thread](https://hypothes-is.slack.com/archives/C074BUPEG/p1678214130862539).

Upon investigation the culprit appears to be a pair of nested subqueries
within the bulk annotations SQL query that're responsible for finding
the list of groups that the audience are members of: [Slack thread](https://hypothes-is.slack.com/archives/C1MA4E9B9/p1678290077051349).

If you replace this group ids subquery with a more sensible one then the
top-level query executes quickly in Metabase.

So fix the Python code to get SQLAlchemy to generate this more efficient
subquery instead.