Add missing pytz requirement
pytz is used by reset_password.py but we don't have a direct dependency
on it, it's only in our requirements.txt file because celery depends on
it, which means that if celery ever removes its pytz dependency our
reset_password.py will break. Add pytz as a direct dependency to fix
this.

https://github.com/hypothesis/h/blob/f98179e4e4ca50142a06803cddc5c694eea0d5a4/h/schemas/forms/accounts/reset_password.py#L53