Create an annotation summary table which summaries counts by various facets
This greatly duplicates a number of other tables we have, and is the result
of requirements becoming clearer over time. Basically we want to count
everything in combination with everything else pretty much.

We probably want to come back round and try and refactor all the other tables
out which could be derived from this table.

The main value add here is the addition of the annotation sub-type.