Increase rate limit for API and asset requests to 4/sec
The original values were chosen very conservatively, but we think this has had
the effect of slowing down bulk annotation operations (eg. via the client's
Import feature [1]) unnecessarily.

The new value for API requests has been chosen based on the 95th percentile
response times of the `GET /api/search` and `POST /api/annotations` endpoints,
which are two endpoints that are both heavily used and expensive. The thinking
is that as long as the significant majority of requests are processed faster
than this, we'll avoid a backlog.

The new value for asset requests was chosen in a similar way, but is much higher
because asset requests are much faster at the 95th percentile.

[1] https://github.com/hypothesis/client/issues/5741#issuecomment-1699245104