Remove support for specifying fields on the bulk annotation API
The service supported any combinations of the allowed fields but the API
locked them two only support both of them in the same order.

Adding new fields (eg the new metadata field) is very easy at the
service level but the limitations at the API level make it more
cumbersome.