Optimize readable groups query
This query is not very expensive in absolute terms but it done very
often.

We'll always query for the world readable groups as a separate query,
this only yields a few hundred groups.

For the case when we use a user.id join on the membership table instead
of doing a subquery. We don't need to join on the user table as the
previous version, just the membership one.

As a last step avoid going over the results to pluck the value, return
them as a list of scalars directly from SQLA.