Remove the order clause to find candidate annotations
The process to find missing annotations in the slim table is
significantly.
slower when we order by `created`.

This change will mean that finish all recent annotations will take
longer but the overall process will be much faster.

We've also seen the behaviour where the first batch of day takes a long
time and the rest are much faster. This is probably due to trying to
load the `created` index in memory.