Filter candidate annotations by dates
This should limit the Annotation.id that need to be compared with
AnnotationSlim.pubid making the query faster.

The default values here avoid having to sync the deploy with the
h-periodic task.

The since value will pick the first annotations in DB and the until will
provide enough annotations (~2M annos) to work on for a few days.