Remove `bin/hypothesis init` and `bin/hypothesis migrate`
Replace `bin/hypothesis init` with a much simpler standalone
`bin/make_db` script similar to what other apps have, and replace
`bin/hypothesis migrate` with just calling `alembic` directly as we do
in other apps.

This is going to make it easier to apply our cookiecutter template
(https://github.com/hypothesis/cookiecutters) to h because it makes the
`make db` recipe in h's `Makefile` the same as the one in the
cookiecutter's `Makefile` template.

The new `bin/make_db` script that this commit adds to h is slightly
different from the cookiecutter's `bin/make_db` template but the same in
spirit and it's easy for h to override the whole `bin/make_db` file with
its own version when using the cookiecutter.

Instead of calling `bin/hypothesis init` and `bin/hypothesis migrate`,
`make db` now just calls a simple `bin/make_db` script and then calls
`alembic` directly.

The new `bin/make_db` script is similar to the one in the cookiecutter
(used for example in Via) and is just a minimal script that bootstraps a
Pyramid application environment (in order to get the app's Pyramid
settings) and then, if the DB is not already stamped by Alembic,
initializes and stamps the DB. Unlike in other apps, h's new
`bin/make_db` also initializes the Elasticsearch index.