Remove `bin/hypothesis shell`
Remove h's `bin/hypothesis shell` command and just call Pyramid's
`pshell` directly instead, like we do in other projects.

This will make it easier to apply our cookiecutter template
(https://github.com/hypothesis/cookiecutters) to h because it makes the
`shell` target in h's `Makefile` the same as the one in the
cookiecutter's `Makefile` template.

It's also just much simpler.

`h/pshell.py` is added to contain our pshell customisations, as in other
projects.

`pyramid-ipython` is added as a dependency so that `make shell` launches
an IPython shell instead of a standard Python shell.