Handle deleted groups while updating the annotation slim table
The process of deleting a group leaves the row in annotation pointing to
a non existing group until the annotations itself is deleted.

The annotation slim uses FKs to other tables so that's not currently
necessary. Skip any action on the upsert annotation slim method and let
the CASCADE on the annotation table remove the existing slim rows if
necessary.