Remove no-longer-needed Heroku support
This originates from:

https://github.com/hypothesis/h/commit/bd6948423a88970200b006316ae6430c0af4e93a

It looks like it was added in order to support Dokku, which was a
service that we used to use to deploy development branches of h to
temporary servers for testing (and apparently these temporary servers
ran on Heroku?).

We haven't used Dokku or Heroku for years.