Validate max length for document.link.href
We need to validate the max length as we are inserting some value in
the DB which exceed the size for an indexable value.

The value selected here, 3000, is longer that the documented max length
supported by some browsers (2083) but that allows for any encoding
differences.

At the time of writing only around 15 ros in the DB out of a
total of 14 milling documents are over 3000 characters long.

The main focus now is having this under control in the storage layer, we
could refine this further but it doesn't seem like a common occurrence
in any case.