Record user deletions in a new DB table
Remove the CLI command for deleting a user: this interacts awkwardly
with recording deletions in a table because there's no authenticated
user to record as the requester of the deletion. There could be various
solutions to this but no one uses this CLI command anyway so let's just
delete it.