Replace arbitrary font-size's and line-height's
In the new create-group form replace the arbitrary `font-size` and
`line-height` values (which were copied from the legacy forms) with
values from Tailwind's default `font-size` and `line-height` scales.

Heading `font-size`: the legacy page uses a `font-size` of `19px`, so
use Tailwind's `text-xl` which gives a `font-size` of `20px` (there's no
default Tailwind class matching `19px` exactly).

Heading `line-height`: the legacy page uses a `line-height` of of `15px`
for the heading, which is smaller than the `font-size` and causes lines
to overlap with each other if a heading wraps onto multiple lines. This
`line-height` may be an accident: the heading gets the `line-height`
from a CSS rule applied to the `<body>` element. Use Tailwind's `/none`
which gives a `line-height` of `1` (which works out as `20px`, the same
as the `font-size`). We could use `/3` instead (which would give a
`line-height` of `12px`), this would be closer to the legacy
`line-height` (`13px`) but I feel that the `line-height` should be at
least equal to the `font-size` otherwise headings look bad when they
wrap onto multiple lines, so I've chosen the smallest `line-height` in
Tailwind's scale that is at least equal to the `font-size`.

Label `font-size`: the legacy page uses `13px` for the form element
labels, button text, and footer. Tailwind's `font-size` scale has no
class that matches `13px` exactly. It has either `text-xs` (`12px`)
which I feel would be too small (the legacy `13px` is already probably
too small, we don't want to go even smaller) or `text-sm` (`14px`). So
use `text-sm` as the closest match that is at least as large as the
legacy `font-size`.

Label `line-height`: the legacy page uses a `15px` (roughly `1.15`)
`line-height` for the form element labels, button text, and footer. Use
Tailwind's `/tight` (`1.25`, `17.5px`) as the closest matching leading
in Tailwind's scale.