Replace arbitrary margins and paddings
In the new create-group form, replace arbitrary margins and paddings
(which were copied from the legacy form) with the nearest matching
default Tailwind classes.

I've arbitrarily chosen to go with the nearest Tailwind classes that are
at least as large as the legacy margins and paddings they're replacing,
not smaller.  An earlier commit did the same with `font-size`'s and
`line-height`'s.