Use `matched_route` not `path`
Use `request.matched_route.name` rather than `request.path` to determine
whether a request is an API request.

This is simpler to implement and it seems better for code to be coupled
to `request.matched_route.name` (an internal value) than to be coupled
to the public `request.path`: it means we can change the public paths of
our endpoints without potentially breaking random parts of our code that
depend on them.

It's not necessary to special-case the `"/api/token"` and `"/api/badge"`
endpoints (non-API endpoints whose paths nonetheless start with
`"/api/"`): the route names for these endpoints do not start with
`"api."` (their route names are just `"token"` and `"badge"`).