Extract `is_api_request()` and add unit tests
Extract `is_api_request()` into a shared helper function (instead of a
private method) and add unit tests.

Future commits will add code in multiple places that call the new shared
helper function.

But even with only one caller for the function there are still benefits
to extracting it and unit-testing it directly: when writing the tests I
noticed that it was returning `None` instead of `False` and added a
`bool()` call to fix this.