Make the "Create group" button actually work
Make the "Create group" button (or hitting Enter when the keyboard focus
is in the Name field) actually call the API to create the group and then
redirect the browser to the newly-created group's page.

Also handle errors from the API by displaying error messages to the
user.

Known issues:

* A min-length should be added to the name field so that it's not
  possible to submit a too-short name that the API will reject. This
  will be added in a future commit.
* A spinner or loading state of some sort needs to be displayed when the
  API request is in flight. This will be added in a future commit.