Rename `combined.py` and `SecurityPolicy`
Rename `SecurityPolicy` to `TopLevelPolicy` and rename the `combined.py`
file that contains it to `top_level.py`.

This is consistent with the naming of other security policies and their
files in the same directory, e.g. `BearerTokenPolicy` and
`bearer_token.py` etc.

Besides the inconsistency, `combined.py` was also a weird name and it
was confusing that the file was named differently than the class it
contains.