Remove `_basic_http_auth.py` to `_auth_client.py`
Make the name of the file consistent with the name of the class that it
contains (`AuthClientPolicy`). This is consistent with how other classes
and their files are named in this directory, e.g. `CookiePolicy` and
`_cookie.py` etc.

Aside from the inconsistency `_basic_http_auth.py` was also a misleading
name: the file does not contain a generic HTTP basic auth policy. The
policy specifically allows request to be authenticated using an auth
client, and happens to use basic auth to transfer the client ID and
secret.