Clear error message when re-submitting
Clear any error message from the creeate-group form when the form is
re-submitted.

It currently looks a bit odd when the form is already showing an error
message and the user re-submits the form and gets the same error message
from the server: nothing changes and the error message just continues to
be shown.

I think it's better if the error message disappears when the form is
re-submitted and re-appears again if another error response is received:
this makes it clear that the action was re-tried and another error was
received.