Remove an unnecessary optimisation
Remove an unnessary optimisation where
`AuthCookieService.verify_cookie()` (a method that is only ever called
by `CookiePolicy.identity()`, and returns the user identity represented
by the request's auth cookie) caches the user object as `self._user` so
that it can just return it if called again, rather than recomputing it
from the cookie.

This optimisation looks potentially dodgy to me because `self._user` is
not only set by the `verify_cookie()` method but also by
`create_cookie()`.

We probably don't need to optimise this but if we do want to optimise it
the right way would be to cache the final return value of the
`AuthCookieService.verify_cookie()` method (not an intermediate value)
or cache the return value of the higher-level `CookiePolicy.identity()`
method that calls it.