Fix Sass deprecation warnings by opting into new behavior
Following a recent Sass update, the CSS build process was producing many
warnings about upcoming Sass changes affecting declarations that appear after
nested rules.  See
https://sass-lang.com/documentation/breaking-changes/mixed-decls/.

One of the approaches suggested on this page to resolve a deprecation is to opt
into the new behavior by wrapping the declarations causing the warning in a
`& { }` block. This is the approach taken in this commit.