Log a message when issuing a missing API cookie
Log a message when we issue an API auth cookie because we received a
request that was authenticated with an HTML auth cookie but which was
missing an API auth cookie.

The log message will enable us to know whether the code for detecting
and fixing this situation (HTML auth cookie but no API one) is necessary
to keep around or not.