Un-pin upload-artifact action and enable `include-hidden-files` option
The `include-hidden-files` setting is required in future releases of this action
to enable uploading files whose names begin with a period.  We pinned the action
previously because this config option was broken when initially added, but it
has now been fixed.

See https://github.com/actions/upload-artifact/releases/tag/v3.2.1-node20.