Remove an unnecessary optimization
Remove an unnecessary optimization where
`AuthTicketService.verify_ticket()` caches the verified ticket's user
as `self._user` so that, if `verify_ticket()` is called again during the
same request, it can just return `self._user` and avoid re-verifying the
ticket (including querying the DB for the ticket).

This is an unnecessary optimization that adds a lot of code complexity
and is making future refactorings unnecessarily difficult. I doubt the
optimization is worth the maintenance difficulty.

`AuthTicketService.verify_ticket()` is only called by
`AuthTicketCookieHelper.identity()` which is only called by
`CookiePolicy.identity()` and `APICookiePolicy.identity()` and only one
of `CookiePolicy` or `APICookiePolicy` will be effective for a given
request. So the only way that `AuthTicketService.verify_ticket()` will
be called multiple times within the same request is if either
`CookiePolicy.identity()` or `APICookiePolicy.identity()` is called
multiple times within the same request. If an optimization was needed
here it would make more sense to do it higher up in
`AuthTicketCookieHelper.identity()` or in `CookiePolicy.identity()` and
`APICookiePolicy.identity()`.