Remove unused group upsert API
This API was added for our own LMS app:

https://github.com/hypothesis/h/pull/5428

But the LMS app doesn't use this API anymore (and actually I'm not sure
it ever did) and it doesn't look like anything else ever used it.

In upcoming PRs we need to make changes to group-related APIs (for
example instead of only a group's creator being able to edit/update a
group, all of a group's owners and admins need to be able to). Let's
first remove this unused API so we don't have to deal with updating it,
especially since this API's code is entangled with other APIs that we do
need to update.

Slack thread: https://hypothes-is.slack.com/archives/C4K6M7P5E/p1724926065440579