Simplify some code
For open groups the group's home page shows the group's creator as its
only member (or none if the group has no creator), whereas for other
types of group it shows the group's list of members.

The page does not show "moderators". It only ever shows a list of
"members" (which may be the group's list of members or may be just the
creator depending on the group type).

It looks like there may have once been an intention to have a list of
group moderators and to show that on the home pages of open groups, but
that never came to pass.

Refactor this code to make the actual current behaviour clearer.